# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DMD Church Management API
  version: 2.0.0
  description: |
    Central API for the DMD church management system.

    ## Authentication

    Two authentication mechanisms are supported:

    ### Session (web users)
    1. `POST /api/auth/login` → receive a signed HMAC session cookie + CSRF token.
    2. Include the cookie automatically on every subsequent request.
    3. Include the `X-CSRF-Token` header (value from login response) on all
       state-changing requests (POST / PATCH / DELETE).

    ### API Key (Telegram bots)
    Include `Authorization: Bearer <BOT_API_KEY>` on every request.
    Bot keys bypass CSRF validation.
servers:
  - url: https://member-evolution-dashboard-xxxx-ew.a.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerApiKey:
      type: http
      scheme: bearer
  schemas:
    OkTrue:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
        errors:
          type: array
          items:
            type: string
...
    Member:
      type: object
      properties:
        id:    { type: string }
        name:  { type: string }
        zone:  { type: string }
        aliases: { type: string }
...
    Student:
      type: object
      properties:
        name:    { type: string }
        class_id:{ type: string }
        subgroup:{ type: string }
        status:  { type: string }
    Lesson:
      type: object
      properties:
        lesson_id:    { type: string }
        lesson_title: { type: string }
        lesson_date:  { type: string, format: date }
        class_id:     { type: string }
    AttendanceEvent:
      type: object
      properties:
        event_id:    { type: string }
        event_name:  { type: string }
        date:        { type: string, format: date }
        description: { type: string }
    AttendanceCategory:
      type: object
      properties:
        category_id:   { type: string }
        category_name: { type: string }

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/bot/lessons:
    post:
      tags: [bot]
      summary: Record a lesson (attendance bot)
      description: |
        Attendance bot submits a parsed lesson block.
        Rate limited: 120 req/min per IP.
      security:
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [classCode, date, title, instructor]
              properties:
                classCode:  { type: string, example: "164-2C" }
                date:       { type: string, format: date }
                title:      { type: string, example: "Introduction à la foi" }
                instructor: { type: string }
                students:
                  type: array
                  items:
                    type: object
                    required: [name, status]
                    properties:
                      name:     { type: string }
                      status:   { type: string, enum: [present, absent, unknown] }
                      subgroup: { type: string }
                      note:     { type: string }
                mode:    { type: string, enum: [create, replace, delete], default: create }
                lessonId:{ type: string }
                classId: { type: string }
                source:  { type: string, example: attendance_bot }
      responses:
        '200':
          description: Lesson recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkTrue'
                  - type: object
                    properties:
                      source:  { type: string }
                      summary: { type: object }
                      result:  { type: object }
        '400':
          description: Missing required field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Firestore not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/bot/meetings:
    post:
      tags: [bot]
      summary: Record a meeting (mannam bot)
      security:
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [summary, date]
              properties:
                summary:         { type: string }
                date:            { type: string, format: date }
                time:            { type: string, example: "14:30" }
                location:        { type: string }
                description:     { type: string }
                participants:
                  type: array
                  items: { type: string }
                calendarEventId: { type: string }
                source:          { type: string, default: mannam_bot }
      responses:
        '200':
          description: Meeting recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkTrue'
                  - type: object
                    properties:
                      meetingId: { type: string }
                      summary:   { type: object }
        '400':
          description: Missing required field
        '401':
          description: Missing or invalid API key
        '503':
          description: Firestore not configured
  /api/bot/meetings/{id}:
    delete:
      tags: [bot]
      summary: Delete a meeting (mannam bot)
      security:
        - bearerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkTrue'
                  - type: object
                    properties:
                      meetingId: { type: string }
        '401':
          description: Missing or invalid API key
        '503':
          description: Firestore not configured
  /api/bot/members:
    get:
      tags: [bot]
      summary: Get member list (both bots)
      description: Used by bots to resolve member names before submitting records.
      security:
        - bearerApiKey: []
      responses:
        '200':
          description: Member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:      { type: boolean }
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
  /api/bot/events:
    get:
      tags: [bot]
      summary: List attendance events
      security:
        - bearerApiKey: []
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:     { type: boolean }
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/AttendanceEvent' }
  /api/bot/categories:
    get:
      tags: [bot]
      summary: List attendance categories
      security:
        - bearerApiKey: []
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:         { type: boolean }
                  categories:
                    type: array
                    items: { $ref: '#/components/schemas/AttendanceCategory' }
  /api/bot/events/{eventName}/attendance:
    get:
      tags: [bot]
      summary: Get participants for an attendance event
      security:
        - bearerApiKey: []
      parameters:
        - name: eventName
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Participants
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:         { type: boolean }
                  event_name: { type: string }
                  rows:
                    type: array
                    items:
                      type: object
                      properties:
                        participant_name: { type: string }
                        category:         { type: string }
                        timestamp:        { type: string }
  /api/bot/events/{eventName}/participants:
    post:
      tags: [bot]
      summary: Add participants to an event
      security:
        - bearerApiKey: []
      parameters:
        - name: eventName
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participants]
              properties:
                participants:
                  type: array
                  items: { type: string }
                category:
                  type: string
                  default: Guest
      responses:
        '200':
          description: Added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:      { type: boolean }
                  added:   { type: array, items: { type: string } }
                  skipped: { type: array, items: { type: string } }
  /api/bot/events/{eventName}/participants/{participantName}:
    delete:
      tags: [bot]
      summary: Remove a participant from an event
      security:
        - bearerApiKey: []
      parameters:
        - name: eventName
          in: path
          required: true
          schema: { type: string }
        - name: participantName
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:               { type: boolean }
                  removed:          { type: boolean }
                  participant_name: { type: string }
  /api/bot/academy/report/class/{code}:
    get:
      tags: [bot]
      summary: Full class attendance report
      security:
        - bearerApiKey: []
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
          example: "164-2C"
      responses:
        '200':
          description: Class report
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:         { type: boolean }
                  cls:        { type: object }
                  lessons:
                    type: array
                    items: { $ref: '#/components/schemas/Lesson' }
                  students:
                    type: array
                    items: { $ref: '#/components/schemas/Student' }
                  att_lookup: { type: object }
        '404':
          description: Class not found
  /api/bot/academy/report/student/{name}:
    get:
      tags: [bot]
      summary: Student attendance history
      security:
        - bearerApiKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
          example: "Marie%20Dupont"
      responses:
        '200':
          description: Student records
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:           { type: boolean }
                  student_name: { type: string }
                  records:
                    type: array
                    items:
                      type: object
                      properties:
                        lesson_title: { type: string }
                        lesson_date:  { type: string }
                        class_code:   { type: string }
                        status:       { type: string }
        '404':
          description: No data for this student
  /api/bot/academy/report/absentees/{code}:
    get:
      tags: [bot]
      summary: Absentees grouped by lesson
      security:
        - bearerApiKey: []
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - name: lesson
          in: query
          required: false
          schema: { type: string }
          description: Optional substring filter on lesson title
      responses:
        '200':
          description: Absentee report
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:           { type: boolean }
                  cls:          { type: object }
                  lessons:      { type: array }
                  att_by_lesson:{ type: object }
        '404':
          description: Class not found